Routes and small refactors
