+++ src/domain/communication/discussion/constants/DiscusionCategories.ts
@@ -5,3 +5,11 @@ export enum DiscussionCategoryExtEnum {
 }
 
 export type DiscussionCategoryExt = DiscussionCategory | DiscussionCategoryExtEnum;
+
+export const ForumCategories = [
+  DiscussionCategory.PlatformFunctionalities,
+  DiscussionCategory.CommunityBuilding,
+  DiscussionCategory.ChallengeCentric,
+  DiscussionCategory.Help,
+  DiscussionCategory.Other,
+];
