+++ src/domain/communication/discussion/routing/ForumRoute.tsx
@@ -2,6 +2,7 @@ import { FC } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import { Error404 } from '../../../../core/pages/Errors/Error404';
 import { nameOfUrl } from '../../../../core/routing/urlParams';
+import DiscussionPage from '../pages/DiscussionPage';
 import ForumPage from '../pages/ForumPage';
 
 interface ForumRouteProps {}
@@ -17,7 +18,7 @@ export const ForumRoute: FC<ForumRouteProps> = () => {
       <Route path={'/'}>
         <Route index element={<ForumPage />} />
         <Route path="/new" element={<ForumPage dialog="new" />} />
-        <Route path={`:${nameOfUrl.discussionId}`} element={<ForumPage />} />
+        <Route path={`discussion/:${nameOfUrl.discussionId}`} element={<DiscussionPage />} />
         <Route path="*" element={<Error404 />} />
       </Route>
     </Routes>
